python/requirements.txt
torch>=2.0
torch-geometric>=2.4
faiss-cpu>=1.7
onnx>=1.14
onnxruntime>=1.15
pyyaml>=6.0
pytest>=7.0
